feat(version): update version to 3.0.13-beta across all relevant files

+++ springboot-backend/src/main/java/com/admin/service/impl/UserTunnelExitPolicyServiceImpl.java
@@ -95,7 +95,7 @@ public class UserTunnelExitPolicyServiceImpl extends ServiceImpl<UserTunnelExitP
             }
             UserTunnelExitPolicy policy = new UserTunnelExitPolicy();
             policy.setUserTunnelId(userTunnelId);
-            policy.setTunnelId(userTunnel.getTunnelId().longValue());
+            policy.setTunnelId(userTunnel.getTunnelId());
             policy.setExitNodeId(exit.getNodeId());
             policy.setFlowQuotaGb(null);
             policy.setUsedFlow(0L);
